fix(processData): I fixed the structure and added the .py that preprocess the data.
